+++ packages/react/src/index.ts
@@ -5,7 +5,6 @@ export * from "./model-context";
 export * from "./primitives";
 export * from "./runtimes";
 export * from "./types";
-export * from "./ui";
 
 export { useToolArgsFieldStatus as unstable_useToolArgsFieldStatus } from "./utils/json/parse-partial-json";