+++ packages/react/src/primitives/index.ts
@@ -9,7 +9,6 @@ export * as ThreadPrimitive from "./thread";
 export * as ThreadListPrimitive from "./threadList";
 export * as ThreadListItemPrimitive from "./threadListItem";
 
-export { useContentPartDisplay } from "./contentPart/useContentPartDisplay";
 export { useContentPartImage } from "./contentPart/useContentPartImage";
 export { useContentPartText } from "./contentPart/useContentPartText";
 export { useContentPartReasoning } from "./contentPart/useContentPartReasoning";