+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @assistant-ui/react
 
+## 0.8.0
+
+### Breaking changes
+
+- pre-styled UI components have moved to `@assistant-ui/react-ui`
+- run `npx assistant-ui upgrade` to update your codebase
+- manual upgrade: change styled component imports from `@assistant-ui/react` to `@assistant-ui/react-ui` (and same with markdown)
++ manual upgrade: change styled component imports from `@assistant-ui/react` to `@assistant-ui/react-ui` (and same with Markdown)
+
 ## 0.7.91
 
 ### Patch Changes